--- src/components/layout-header.js
@@ -1,9 +1,0 @@
-import Nav from './nav'
-
-export default function LayoutHeader() {
-	return (
-		<header>
-			<Nav/>
-		</header>
-	)
-}
