+++ src/pages/settings/account.js
@@ -1,6 +1,6 @@
-import {useNavigate} from 'react-router-dom'
+import {Link, useNavigate} from 'react-router-dom'
+import LayoutSettings from '../../layouts/settings'
 import DeleteUserForm from '../../components/delete-user-form.js'
-import {Link} from 'react-router-dom'
 
 export default function Account({dbSession}) {
 	const {session} = dbSession
@@ -9,23 +9,17 @@ export default function Account({dbSession}) {
 	const handleDeleteUser = () => navigate('/logout')
 
 	return (
-		<main>
-			<header>
-				<h1>
-					Settings
-				</h1>
-				<nav>
-					<Link to="/settings/account">Account</Link>
-					<Link to="/settings/channels">Channels.</Link>
-				</nav>
-			</header>
+		<LayoutSettings>
 			<p>
-				This is your Radio4000 account: <strong>{session?.user?.email}</strong>
+				Email: <kbd>{session?.user?.email}</kbd>
 			</p>
-			<p>Change your password? <Link to="/reset-password">Change</Link></p>
 			<p>
-				Delete your account <DeleteUserForm onDelete={handleDeleteUser} />
+				Password: <Link to="/reset-password">change your password?</Link>
 			</p>
-		</main>
+			<p>
+				Delete your account:
+			</p>
+			<DeleteUserForm onDelete={handleDeleteUser}/>
+		</LayoutSettings>
 	)
 }
